BL-414 #resolve
Renaming of Box Cache functions to standardized `cache{function}()`
