+++ src/main/java/ortus/boxlang/runtime/bifs/global/cache/Cache.java
@@ -27,14 +27,14 @@ import ortus.boxlang.runtime.types.Argument;
 import ortus.boxlang.runtime.validation.Validator;
 
 @BoxBIF
-public class GetBoxCache extends BIF {
+public class Cache extends BIF {
 
 	private static final Validator cacheExistsValidator = new CacheExistsValidator();
 
 	/**
 	 * Constructor
 	 */
-	public GetBoxCache() {
+	public Cache() {
 		super();
 		declaredArguments = new Argument[] {
 		    new Argument( false, Argument.STRING, Key.cacheName, Key._DEFAULT, Set.of( cacheExistsValidator ) )
