+++ src/main/java/ortus/boxlang/runtime/bifs/global/cache/CacheService.java
@@ -18,28 +18,26 @@ import ortus.boxlang.runtime.bifs.BIF;
 import ortus.boxlang.runtime.bifs.BoxBIF;
 import ortus.boxlang.runtime.context.IBoxContext;
 import ortus.boxlang.runtime.scopes.ArgumentsScope;
-import ortus.boxlang.runtime.types.Array;
 
 @BoxBIF
-public class GetBoxCacheNames extends BIF {
+public class CacheService extends BIF {
 
 	/**
 	 * Constructor
 	 */
-	public GetBoxCacheNames() {
+	public CacheService() {
 		super();
 	}
 
 	/**
-	 * Get an array of all the registered caches in BoxLang
+	 * Get access to the Box Cache services
 	 *
 	 * @param context   The context in which the BIF is being invoked.
 	 * @param arguments Argument scope for the BIF.
 	 *
-	 * @return An array of all the registered caches in BoxLang
+	 * @return The cache service instance
 	 */
-	public Array _invoke( IBoxContext context, ArgumentsScope arguments ) {
-		// Get the requested cache
-		return this.cacheService.getRegisteredCaches();
+	public ortus.boxlang.runtime.services.CacheService _invoke( IBoxContext context, ArgumentsScope arguments ) {
+		return this.cacheService;
 	}
 }
