+++ src/test/java/ortus/boxlang/runtime/interop/DynamicInteropServiceTest.java
@@ -789,7 +789,7 @@ public class DynamicInteropServiceTest {
 				"""
 					import java.util.Arrays;
 
-					cacheService = getBoxCacheService()
+					cacheService = cacheService()
 					cacheService.createDefaultCache( "bddTest" );
 
 					cache = cacheService.getCache( "bddTest" );
